move function from compile_reverse to code, make more functions private
